Revert 9a3d807a7530ea898e852192d65cbe40bab45924 which attempted to escape translsations before applying them to the source code.
In hindsight, we must rely on the translations being correct as we will not know how a certain translation needs to be encoded. The code above was causing some code to be double escaped as if the text to be translated was already escaped and the translator did the same with the translation (as they should) then this plugin would escape it a second time, often breaking the source code.

This acknowledges that the translations supplied by loco CAN break the source code. It is very important that translators use the same speechmarks, as is shown in the source text otherwise we may have problems.